feat: rename Badge.tsx to badge.tsx
